flutterapp backend/requirements.txt
Flask
flask-cors
flask-pymongo
python-dotenv
bcrypt
PyJWT
geopy